==> R_scripts/4_Fig2_gases_by_location.R <==
## New and improved script to make a new gas figure that includes everything
##
## 2022-04-22 (updated 6/22/23)
## Peter Regier
##
# ############ #
# ############ #

# 1. Setup ---------------------------------------------------------------------

## Load packages
require(pacman)
p_load(tidyverse, #keep things tidy
       cowplot, #plot_grid
       ggpubr, #stat_compare_means()
       ggpattern, #geom_boxplot_pattern()
       rstatix, #levene_test()
       janitor,
       ggConvexHull,
       confintr,
       ggallin) #pseudolog10_trans()

## Set ggplot theme
theme_set(theme_bw())

## Set color theme
color_theme <- c("#1B264F", "#CE8147", "#FAF33E", "#3BB273")

## Set comparisons for stats
compare_transect <- list( c("Sediment", "Wetland"), c("Sediment", "Transition"), 
                          c("Sediment", "Upland"), c("Wetland", "Transition"), 
                          c("Wetland", "Upland"), c("Transition", "Upland"))

## Make it easier to reference ID columns
common_cols <- c("kit_id", "transect_location")


# 2. Import and merge datasets -------------------------------------------------

## Read in full dataset
df <- read_csv("data/230623_master_data.csv") %>% 
  select(all_of(common_cols), contains("uM_hr")) %>% 
  mutate(transect_location = str_to_sentence(transect_location)) %>% 
mutate(transect_location = fct_relevel(transect_location, c("Sediment", "Wetland", "Transition", "Upland")))

## Also read in seawater concentrations and calculate means
sw <- read_csv("data/230623_ghg_saltwater.csv") %>% 
  select(contains("uM_hr")) %>% 
  summarize(across(everything(), mean))


# 3. Make figures and export ---------------------------------------------------

make_boxplot_fig2 <- function(var, y_label, y_max){
  
  df_trim <- df %>% filter({{var}} < y_max)
  
  ggplot(df_trim, aes(transect_location, {{var}}, fill = transect_location)) + 
    geom_boxplot(show.legend = F, outlier.alpha = 0, width = 0.5, alpha = 0.8) +
    geom_jitter(color = "black", show.legend = F, width = 0.05) + 
    scale_fill_manual(values = color_theme) + 
    labs(x = "\n Location along inundation gradient", y = y_label) + 
    stat_compare_means(comparisons = compare_transect, label = "p.signif")
  }

make_boxplot_supp <- function(var, y_label){
  ggplot(df, aes(transect_location, {{var}}, fill = transect_location)) + 
    geom_boxplot(show.legend = F, outlier.alpha = 0, width = 0.5, alpha = 0.8) +
    geom_jitter(color = "black", show.legend = F, width = 0.05) + 
    scale_fill_manual(values = color_theme) + 
    labs(x = "\n Location along inundation gradient", y = y_label) + 
    scale_y_continuous() + 
  stat_compare_means(comparisons = compare_transect, label = "p.signif")
}

## Make Fig 2 sub-plots
p_do <- make_boxplot_fig2(do_uM_hr, "DO consumption (µM/hr)", 590)
p_co2 <- make_boxplot_fig2(co2_uM_hr, "CO2 production (µM/hr)", 100) + 
  geom_hline(yintercept = sw$co2_uM_hr, linetype = "dashed")
p_ch4 <- make_boxplot_fig2(ch4_uM_hr, "CH4 production (µM/hr)", 0.31) + 
  geom_hline(yintercept = sw$ch4_uM_hr, linetype = "dashed")
p_n2o <- make_boxplot_fig2(n2o_uM_hr, "N2O production (µM/hr)", 0.123) + 
  geom_hline(yintercept = sw$n2o_uM_hr, linetype = "dashed")

## Make Supp fig sub-plots
p_do_s <- make_boxplot_supp(do_uM_hr, "DO consumption (µM/hr)")
p_co2_s <- make_boxplot_supp(co2_uM_hr, "CO2 production (µM/hr)") + 
  geom_hline(yintercept = sw$co2_uM_hr, linetype = "dashed")
p_ch4_s <- make_boxplot_supp(ch4_uM_hr, "CH4 production (µM/hr)") + 
  geom_hline(yintercept = sw$ch4_uM_hr, linetype = "dashed")
p_n2o_s <- make_boxplot_supp(n2o_uM_hr, "N2O production (µM/hr)") + 
  geom_hline(yintercept = sw$n2o_uM_hr, linetype = "dashed")

plot_grid(p_do, p_co2, p_ch4, p_n2o, nrow = 1, labels = c("A", "B", "C", "D"), align = "hv")
ggsave("figures/2_Fig2_gases_by_location.png", width = 13, height = 4.5)
ggsave("figures/2_Fig2_gases_by_location.pdf", width = 13, height = 4.5)


fig_s1 <- plot_grid(p_do_s, p_co2_s, p_ch4_s, p_n2o_s, nrow = 1, labels = c("A", "B", "C", "D"), align = "hv")
ggsave(fig_s1, "figures/S1_gases_by_location_all_data.png", width = 13, height = 4.5)

# 4. Calculate statistics (mean and standard error) for Figure 2 ---------------

## Set up a helper function to paste mean and se together
n_sigfigs = 3
se <- function(var){paste0(round(mean({{var}}), n_sigfigs), 
                          "±", 
                          round(sd({{var}}) / sqrt(length({{var}})), n_sigfigs))}

mean_range <- function(var){paste0(signif(mean({{var}}), n_sigfigs), 
                                   " (",
                                   signif(min({{var}}), n_sigfigs), 
                                   " - ", 
                                   signif(max({{var}}), n_sigfigs), 
                                   ")")}

## Broken out by location
transect_stats_by_location <- df %>% 
  select(transect_location, contains("_uM_hr")) %>% 
  group_by(transect_location) %>% 
  summarize(across(is.numeric, mean_range))
  
## Across all locations
transect_stats <- df %>% 
  select(contains("_uM_hr")) %>% 
  summarize(across(is.numeric, mean_range)) %>% 
  mutate(transect_location = "All locations") %>% 
  relocate(transect_location)

sw_stats <- read_csv("data/230623_ghg_saltwater.csv") %>% 
  select(contains("uM_hr")) %>% 
  summarize(across(is.numeric, mean_range)) %>% 
  mutate(do_uM_hr = NA) %>% 
  mutate(transect_location = "Saltwater") %>% 
  relocate(transect_location)

fig2_stats <- bind_rows(transect_stats,
                        transect_stats_by_location, 
                        sw_stats)

write_csv(fig2_stats, "data/230920_Table_S1.csv")


## Calculate Wilcoxon tests

calculate_wilcoxon <- function(var){
  f = as.formula(paste0(var, "~transect_location"))
  
  wilcox_test(f, data = df) %>% 
    select(contains("group"), p) %>% 
  rename("Transect location 1" = 1, 
         "Transect location 2" = 2, 
         !!var := 3)
}

# Function to format numeric values based on specified conditions
format_numeric_values <- function(x) {
  case_when(
    x < 0.001 ~ "***",
    x < 0.01 ~ "**",
    x < 0.05 ~ "*",
    TRUE ~ as.character(x)
  )
}

# Use rename_with to reformat column names containing "uM_hr"
df <- df 

table_s2 <- bind_cols(calculate_wilcoxon("do_uM_hr"), 
          calculate_wilcoxon("co2_uM_hr") %>% select(-contains("location")), 
          calculate_wilcoxon("ch4_uM_hr") %>% select(-contains("location")), 
          calculate_wilcoxon("n2o_uM_hr") %>% select(-contains("location"))) %>% 
  mutate(across(where(is.numeric), format_numeric_values)) %>% 
  rename_with(
    ~str_to_upper(.x),
    .cols = contains("uM_hr")
  ) %>%
  rename_with(
    ~str_replace_all(.x, "_UM_HR", " (uM/hr)"),
    .cols = contains("_UM_HR")
  )
  
  
write_csv(table_s2, "data/230929_Table_S2.csv")
